requirements.txt
Flask==3.0
requests==2.25.1
gunicorn==21.2.0
python-dotenv